v03
The code has been updated to be much more readable and iterable for future iterations. There are also a few new features that were used a lot in the assessing of the data. You will likely notice that all of the different categories of data (testing, not testing, anomalies, not anomalies, normalized, not normalized, have all have seperate CSV files created when you run the program. this will help run multiple models simultaneously in the future. In addition a silhouette score function has been implemented as the primary method of evaluation for the clustering algorithm for now. The results at this time do not look very good, but there may be a little bit of user error and/or the model needs more tuning at this point.